Refactored hardware monitoring and moved interfaces out to models/
