+++ src/index.ts
@@ -11,7 +11,7 @@ const connectionTimeout =      parseParam(process.env.CONNECTION_TIMEOUT, 30000)
 const maxConcurrentSessions =  parseParam(process.env.MAX_CONCURRENT_SESSIONS, 10);
 const maxQueueLength =         parseParam(process.env.MAX_QUEUE_LENGTH, 10);
 const port =                   parseParam(process.env.PORT, 8080);
-const prebootChrome =          parseParam(process.env.PREBOOT_CHROME, false);
+const prebootChrome =          parseParam(process.env.PREBOOT_CHROME, true);
 const demoMode =               parseParam(process.env.DEMO_MODE, false);
 const enableDebugger =         parseParam(process.env.ENABLE_DEBUGGER, true);
 const maxMemory =              parseParam(process.env.MAX_MEMOMORY_PERCENT, 99);
