+++ src/models/browserless-stats.interface.ts
@@ -1,0 +1,10 @@
+interface IBrowserlessStats {
+  date: number | null;
+  successful: number;
+  error: number;
+  queued: number;
+  rejected: number;
+  memory: number;
+  cpu: number;
+  timedout: number;
+};
