+++ src/models/file-request.interface.ts
@@ -1,0 +1,5 @@
+import * as express from 'express';
+
+export interface IFileRequest extends express.Request {
+  file: any
+}
